# services/control-plane/internal/api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Velane Control Plane API",
    "version": "1.0.0",
    "description": "HTTP API for Velane authentication, tenant administration, workflows, integrations, invocation, observability, and embeds. Unless stated otherwise, authenticated operations accept either a Velane API key (`vl_...`), a session JWT, or the session cookie. The `x-required-scope` extension records the minimum API-key scope or session role."
  },
  "servers": [
    { "url": "/", "description": "Current Velane deployment" }
  ],
...
  "security": [
    { "bearerAuth": [] },
    { "sessionCookie": [] }
  ],
  "paths": {
...
    "/v1/tenant/connections": {
      "get": { "operationId": "listTenantConnections", "summary": "List integration connections", "tags": ["Connections"], "x-required-scope": "invoke", "parameters": [{ "$ref": "#/components/parameters/Search" }, { "$ref": "#/components/parameters/Limit100" }, { "$ref": "#/components/parameters/Offset" }], "responses": { "200": { "description": "Connections", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Connection" } } } } } } },
      "post": { "operationId": "recordConnection", "summary": "Record a completed integration connection", "tags": ["Connections"], "x-required-scope": "manage", "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RecordConnectionRequest" } } } }, "responses": { "201": { "description": "Connection recorded", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Connection" } } } }, "400": { "$ref": "#/components/responses/BadRequest" } } }
    },
...
    "/v1/tenant/connections/{provider}": {
      "parameters": [{ "$ref": "#/components/parameters/Provider" }],
      "delete": { "operationId": "disconnectProvider", "summary": "Disconnect an integration provider", "tags": ["Connections"], "x-required-scope": "manage", "responses": { "204": { "$ref": "#/components/responses/NoContent" }, "404": { "$ref": "#/components/responses/NotFound" } } }
    },
    "/v1/connections": {
      "get": { "operationId": "listConnections", "summary": "List connections for the authenticated token", "tags": ["Connections"], "x-required-scope": "invoke", "parameters": [{ "$ref": "#/components/parameters/Search" }, { "$ref": "#/components/parameters/Limit100" }, { "$ref": "#/components/parameters/Offset" }], "responses": { "200": { "description": "Connections", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Connection" } } } } } } }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "description": "Velane API key (`vl_...`) or RS256 session JWT." },
      "sessionBearer": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT", "description": "RS256 admin session JWT. API keys are not accepted by session-only endpoints." },
      "sessionCookie": { "type": "apiKey", "in": "cookie", "name": "velane_session", "description": "HttpOnly admin session cookie." },
      "embedToken": { "type": "http", "scheme": "bearer", "description": "Opaque embed token (`et_...`). Embed tokens never receive admin scope." },
      "gitWebhookSignature": { "type": "apiKey", "in": "header", "name": "X-Hub-Signature-256", "description": "HMAC-SHA256 signature in `sha256=<hex>` format." },
      "nangoWebhookSignature": { "type": "apiKey", "in": "header", "name": "X-Nango-Hmac-Sha256", "description": "Nango webhook HMAC signature." }
    },
    "parameters": {
...
      "Provider": { "name": "provider", "in": "path", "required": true, "schema": { "type": "string" } },
...
      "Search": { "name": "q", "in": "query", "description": "Case-insensitive search text.", "schema": { "type": "string" } },
      "Limit100": { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100 } },
      "Offset": { "name": "offset", "in": "query", "schema": { "type": "integer", "minimum": 0, "default": 0 } },
...
    },
...
    "responses": {
      "NoContent": { "description": "Operation completed with no response body." },
      "BadRequest": { "description": "Invalid request.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
      "NotFound": { "description": "The requested resource was not found.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
    },
    "schemas": {
      "Error": { "type": "object", "required": ["error"], "properties": { "error": { "type": "string" } } },
...
      "Connection": { "type": "object", "required": ["id", "tenant_id", "provider", "alias", "provider_config_key", "nango_connection_id", "created_at", "updated_at"], "properties": { "id": { "type": "string" }, "tenant_id": { "type": "string" }, "provider": { "type": "string" }, "alias": { "type": "string" }, "provider_config_key": { "type": "string" }, "credential_profile_id": { "type": ["string", "null"] }, "nango_connection_id": { "type": "string" }, "display_name": { "type": "string" }, "created_at": { "type": "string", "format": "date-time" }, "updated_at": { "type": "string", "format": "date-time" } } },
...
      "RecordConnectionRequest": { "type": "object", "required": ["provider"], "properties": { "provider": { "type": "string" }, "display_name": { "type": "string" }, "alias": { "type": "string", "default": "default" }, "credential_profile_id": { "type": "string" } } },
...
    }
  }
}